+++ ChangeLog.md
@@ -8,6 +8,7 @@
 - NEW : Add Button to update rules from proposal lines (MAIN_FEATURE_LEVEL 2)
   To finish this feature, on button click, need to show all modifications applied before validation
 - FIX : Quantity limitation
+- FIX : Security access of card and list
 
 ## 2.2.0 - 13/10/2020